+++ test-core.ini
@@ -128,7 +128,7 @@ handlers = console
 [logger_ckan]
 qualname = ckan
 handlers =
-level = INFO
+level = WARN
 
 [logger_sqlalchemy]
 handlers =
